types: CollectedCoins
